Move service functions to their own file
